Home, about, portfolio, contact, navbar finished
This update adds:

Home component of a React app. Specifically, it includes the use of the Typed.js library to create a typewriter effect for a dynamic text element, as well as the ScrollReveal library to add scrolling animations to different elements on the page. It also includes a function to copy an email address to the clipboard when a button is clicked.

About component in a React application, includes a new handleLinkClick function that smooth scrolls to a specific section of the page, and a new button link to the portfolio page. The component also has improved styling and new content that provides more details about the developer's journey with coding.

Portfolio component to the React app, which displays a section showcasing several projects. Each project is presented with an image, a brief description, and a link to the project's GitHub repository. The component also includes CSS files for styling and media queries.

Added emailjs library to handle sending email from contact form, included form validation and implemented logic to copy email address to clipboard on button click. Also included social links for LinkedIn, GitHub, and Facebook.

Built responsive navbar with smooth scrolling functionality. It allows users to navigate to different sections of a website using links in the navbar, and highlights the currently active section as the user scrolls through the page.